+++ Sources/Interface/WBTextFieldCell.m
@@ -65,7 +65,7 @@ NSRect _adjustTextFrame(WBTextFieldCell *self, NSRect frame) {
   if (!self->wb_tfFlags.middle) return frame;
 
   NSFont *font = [self font];
-  NSInteger offset = floor((NSHeight(frame) - ([font ascender] - [font descender])) / 2);
+  NSInteger offset = (NSInteger)floor((NSHeight(frame) - ([font ascender] - [font descender])) / 2);
   return NSInsetRect(frame, 0.0, offset);
 }
